Plugged in tera; WIP: doing the same for all the table subfields now
