+++ src/parser/types/database/database.sql
@@ -1,0 +1,11 @@
+--
+-- Current Database: {{ name }}
+--
+
+CREATE DATABASE IF NOT EXISTS {{ name }} {% for option in options %}{% if loop.first %}DEFAULT {% endif %}{{ option }} {% endfor %};
+
+USE `{{ name }}`;
+
+{% for table in tables %}
+{{ table }}
+{% endfor %}
